Sync: terrain analysis tools, contours, TA mask, and refined OSM casings
Covers the port tree through f8f9dbe4. New in the published subset:

- fvkit/analysis: profile, path, measure and viewshed, reachable from
  Python via pyfvw, with the analysis toolbar in the app layer.
- Terrain contours from the elevation data rather than the grid, plus
  elevation tiling and the terrain-avoidance mask overlay.
- Path shaping in the canvas, and OSM style work: golf holes, the marsh
  tuft taken from S-52, and casings for minor roads, paths and cycleways.

Pippin stays held back.

+++ port/include/fvkit/canvas/canvas.h
@@ -34,8 +34,12 @@ struct FvColor {
 
 struct Pen {
   FvColor color;
-  int width = 1;          // pixels; stamped square nib (non-AA)
-  std::vector<int> dash;  // on/off run lengths in px; empty = solid
+  int width = 1;  // pixels; stamped square nib (non-AA)
+  // On/off run lengths in pixels; empty = solid. Real numbers, not whole
+  // pixels: a style that derives its runs from the line width (an OSM
+  // line-dasharray is in width units) rounds to a different pattern for every
+  // width, and two widths of the same line must dash alike to case each other.
+  std::vector<double> dash;
 };
 
 struct Brush {
